updated user models

+++ src/modules/users/users.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  Injectable,
+  ConflictException,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UsersRepository } from './repositories/user.repository';
@@ -7,23 +11,53 @@ import { UsersRepository } from './repositories/user.repository';
 export class UsersService {
   constructor(private usersRepository: UsersRepository) {}
 
-  create(createUserDto: CreateUserDto) {
-    return this.usersRepository.create(createUserDto);
+  async create(createUserDto: CreateUserDto) {
+    const findUsername = await this.usersRepository.findByUsername(
+      createUserDto.username,
+    );
+
+    if (findUsername) {
+      throw new ConflictException('Nome de usuário já existe.');
+    }
+
+    const findEmail = await this.usersRepository.findByEmail(
+      createUserDto.email,
+    );
+
+    if (findEmail) {
+      throw new ConflictException('Email já cadastrado.');
+    }
+
+    const user = await this.usersRepository.create(createUserDto);
+
+    return user;
   }
 
-  findAll() {
+  async findAll() {
     return this.usersRepository.findAll();
   }
 
-  findOne(id: string) {
+  async findOne(id: string) {
+    const findUser = await this.usersRepository.findOne(id);
+    if (!findUser) {
+      throw new NotFoundException('Usuário não encontrado.');
+    }
     return this.usersRepository.findOne(id);
   }
 
-  update(id: string, updateUserDto: UpdateUserDto) {
+  async update(id: string, updateUserDto: UpdateUserDto) {
+    const findUser = await this.usersRepository.findOne(id);
+    if (!findUser) {
+      throw new NotFoundException('Usuário não encontrado.');
+    }
     return this.usersRepository.update(id, updateUserDto);
   }
 
-  remove(id: string) {
+  async remove(id: string) {
+    const findUser = await this.usersRepository.findOne(id);
+    if (!findUser) {
+      throw new NotFoundException('Usuário não encontrado.');
+    }
     return this.usersRepository.delete(id);
   }
 }
